+++ test/libsolidity/smtCheckerTests/modifiers/modifier_multi_functions.sol
@@ -22,5 +22,4 @@ contract C
 	}
 }
 // ----
-// Warning: (86-93): Condition is always true.
 // Warning: (311-324): Assertion violation happens here